store seems ok
