+++ interactive_engine/executor/store/src/db/api/mod.rs
@@ -1,44 +1,53 @@
 #![allow(dead_code)]
 
+use crate::db::proto::common::DataLoadTargetPb;
+use crate::db::proto::model::{EdgeIdPb, EdgeKindPb};
+
+pub use self::config::*;
+pub use self::entity::*;
+pub use self::error::*;
+pub use self::graph::*;
+pub use self::old_condition::*;
+pub use self::property::*;
+pub use self::schema::*;
+
 #[macro_use]
 pub mod error;
 pub mod property;
 mod entity;
 mod graph;
-mod condition;
+mod old_condition;
 mod schema;
 mod config;
-
-pub use self::error::*;
-pub use self::entity::*;
-pub use self::condition::*;
-pub use self::config::*;
-pub use self::graph::*;
-pub use self::property::*;
-pub use self::schema::*;
-use crate::db::proto::model::{EdgeKindPb, EdgeIdPb};
-use crate::db::proto::common::DataLoadTargetPb;
-use crate::v2::api::EdgeRelation;
+pub mod multi_version_graph;
+pub mod partition_snapshot;
+pub mod types;
+pub mod condition;
+pub mod partition_graph;
 
 pub type SnapshotId = i64;
 pub type VertexId = i64;
 pub type LabelId = i32;
-pub type PropId = i32;
+pub type PropertyId = i32;
 pub type BackupId = i32;
+pub type EdgeInnerId = i64;
+pub type SerialId = u32;
 pub type GraphResult<T> = Result<T, GraphError>;
+pub type Records<T> = Box<dyn Iterator<Item=GraphResult<T>> + Send>;
 
 pub const MAX_SI: SnapshotId = SnapshotId::max_value() - 1;
 pub const INFINITE_SI: SnapshotId = SnapshotId::max_value();
 
+#[repr(C)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub struct EdgeId {
     pub src_id: VertexId,
     pub dst_id: VertexId,
-    pub inner_id: i64,
+    pub inner_id: EdgeInnerId,
 }
 
 impl EdgeId {
-    pub fn new(src_id: VertexId, dst_id: VertexId, inner_id: i64) -> Self {
+    pub fn new(src_id: VertexId, dst_id: VertexId, inner_id: EdgeInnerId) -> Self {
         EdgeId {
             src_id,
             dst_id,
@@ -49,18 +58,21 @@ impl EdgeId {
     pub fn from_proto(proto: &EdgeIdPb) -> Self {
         Self::new(proto.get_srcId().get_id(), proto.get_dstId().get_id(), proto.get_id())
     }
-}
 
-impl From<crate::v2::api::EdgeId> for EdgeId {
-    fn from(from_edge_id: crate::v2::api::EdgeId) -> Self {
-        EdgeId::new(
-            from_edge_id.get_edge_inner_id() as i64,
-            from_edge_id.get_src_vertex_id() as VertexId,
-            from_edge_id.get_dst_vertex_id() as VertexId,
-        )
+    pub fn get_src_vertex_id(&self) -> VertexId {
+        self.src_id
+    }
+
+    pub fn get_dst_vertex_id(&self) -> VertexId {
+        self.dst_id
+    }
+
+    pub fn get_edge_inner_id(&self) -> EdgeInnerId {
+        self.inner_id
     }
 }
 
+#[repr(C)]
 #[derive(Clone, PartialEq, Eq, Hash)]
 pub struct EdgeKind {
     pub edge_label_id: LabelId,
@@ -77,6 +89,16 @@ impl EdgeKind {
         }
     }
 
+    pub fn get_edge_label_id(&self) -> LabelId {
+        self.edge_label_id
+    }
+    pub fn get_src_vertex_label_id(&self) -> LabelId {
+        self.src_vertex_label_id
+    }
+    pub fn get_dst_vertex_label_id(&self) -> LabelId {
+        self.dst_vertex_label_id
+    }
+
     pub fn from_proto(proto: &EdgeKindPb) -> Self {
         Self::new(proto.get_edgeLabelId().get_id(), proto.get_srcVertexLabelId().get_id(), proto.get_dstVertexLabelId().get_id())
     }
@@ -90,8 +112,8 @@ impl EdgeKind {
     }
 }
 
-impl From<&EdgeRelation> for EdgeKind {
-    fn from(relation: &EdgeRelation) -> Self {
+impl From<&EdgeKind> for EdgeKind {
+    fn from(relation: &EdgeKind) -> Self {
         EdgeKind::new(relation.get_edge_label_id() as LabelId, relation.get_src_vertex_label_id() as LabelId, relation.get_dst_vertex_label_id() as LabelId)
     }
 }
