+++ interactive_engine/executor/store/src/db/graph/bench/graph.rs
@@ -114,7 +114,7 @@ mod benchmark {
             println!("========================================");
         }
 
-        fn bench(&self, label: LabelId, properties: &HashMap<PropId, Value>, test_count: usize) -> f64 {
+        fn bench(&self, label: LabelId, properties: &HashMap<PropertyId, Value>, test_count: usize) -> f64 {
             // warmup
             for _ in 0..test_count {
                 let id = random();
@@ -239,7 +239,7 @@ mod benchmark {
             println!("========================================");
         }
 
-        fn bench(&self, edge_type: &EdgeKind, properties: &HashMap<PropId, Value>, test_count: usize) -> f64 {
+        fn bench(&self, edge_type: &EdgeKind, properties: &HashMap<PropertyId, Value>, test_count: usize) -> f64 {
             // warmup
             self.insert(edge_type, properties, test_count);
 
@@ -249,7 +249,7 @@ mod benchmark {
             test_count as f64 / cost
         }
 
-        fn insert(&self, edge_type: &EdgeKind, properties: &HashMap<PropId, Value>, mut test_count: usize) {
+        fn insert(&self, edge_type: &EdgeKind, properties: &HashMap<PropertyId, Value>, mut test_count: usize) {
             let mut inner_id = 0;
             while test_count > 0 {
                 let src_id = random();
