+++ interactive_engine/executor/store/src/db/graph/bench/types.rs
@@ -21,7 +21,7 @@ pub fn multi_numeric_properties_type_def(count: usize) -> TypeDef {
     let types = vec![ValueType::Int, ValueType::Long, ValueType::Float, ValueType::Double];
     for i in 0..count {
         let r#type = types[i % types.len()];
-        builder.add_property(i as PropId + 1, i as PropId + 1, "multi_num".to_string(), r#type, None, false, "cmt".to_string());
+        builder.add_property(i as PropertyId + 1, i as PropertyId + 1, "multi_num".to_string(), r#type, None, false, "cmt".to_string());
     }
     builder.build()
 }
@@ -30,7 +30,7 @@ pub fn multi_string_properties_type_def(count: usize) -> TypeDef {
     let mut builder = TypeDefBuilder::new();
     builder.version(1);
     for i in 0..count {
-        builder.add_property(i as PropId + 1, i as PropId + 1, "multi_str".to_string(), ValueType::String, None, false, "cmt".to_string());
+        builder.add_property(i as PropertyId + 1, i as PropertyId + 1, "multi_str".to_string(), ValueType::String, None, false, "cmt".to_string());
     }
     builder.build()
 }
