+++ interactive_engine/executor/store/src/db/graph/tests/backup.rs
@@ -10,7 +10,7 @@ pub fn test_backup_engine<G: GraphStorage>(graph: G, test_dir: &str) {
     // insert data
     let schema_version = 1;
     graph.create_vertex_type(10, schema_version, 1, &types::create_test_type_def(1), schema_version).unwrap();
-    let properties: HashMap<PropId, Value> = HashMap::new();
+    let properties: HashMap<PropertyId, Value> = HashMap::new();
     graph.insert_overwrite_vertex(11, 1, 1, &properties).unwrap();
     graph.insert_overwrite_vertex(12, 2, 1, &properties).unwrap();
