+++ interactive_engine/executor/store/src/db/mod.rs
@@ -6,3 +6,4 @@ pub mod graph;
 mod util;
 #[allow(bare_trait_objects)]
 pub mod proto;
+pub mod wrapper;
