--- interactive_engine/executor/store/src/v2/graph/mod.rs
@@ -1,3 +1,0 @@
-mod rocks_graph;
-mod iter;
-mod entity;
